pyproject.toml
[tool.poetry]
name = "morpheus-hypnos"
version = "0.1.0"
description = "WEB Hypnos automated tests (ZEUS OS with Fiber-v4)."
authors = ["Pedro Thaines <pedro.thaines@intelbras.com.br>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
robotframework = "^6.1.1"
robotframework-browser = "^17.5.1"
robotframework-requests = "^0.9.5"
pillow = "^10.0.1"
imagehash = "^4.3.1"
black = "^23.9.1"
robotframework-tidy = "^4.5.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
